sui-dockerfile: property-prove assembly-agnostic content-addressing + D6 fall-through safety; fix two real bugs the tests surfaced
Layers property + safety tests onto the Phase 1-4 dockerfile pipeline and
fixes two shipped bugs the new tests exposed.

Tests added (proptest, 256 cases each):

sui-spec/tests/dockerfile_content_addressing_properties.rs (8 tests)
  - assembly-agnostic (core thesis): a layer's BLAKE3 key is INVARIANT
    to comment/blank-line noise and to an unconsumed build-arg's value
    (any node strictly before the ARG line), and SENSITIVE to content
    (base image change → every key differs; one instruction change →
    that key differs).
  - partial-invalidation: a 1-instruction change invalidates EXACTLY the
    contiguous downstream suffix; every upstream node stays cache-valid.
    Generalizes the nonroot-gateway 8/11 observation into a property
    over a random change position, plus an anchored fixture cross-check.

sui-dockerfile-wrapper/tests/fall_through_safety_properties.rs (6 tests)
  - D6: for EVERY cache-side failure mode (cache miss, backend
    unreachable, spec the scoped hasher rejects, partial store, absent
    node-cache daemon) the wrapper falls through to exactly one full real
    docker build via the MockCommandRunner and never a pull/splice/
    CacheHit. One control test guards the assertion against being
    vacuously true.

Two real bugs found + fixed at source (tier-honest):

  1. D6 gap: a cache backend error OR a Dockerfile the scoped graph
     hasher rejects (HEALTHCHECK etc., which real docker builds fine) was
     propagated as a WrapperError instead of degrading to a plain build —
     the cache was gating correctness. Refactored to consult_cache(); any
     cache-side error now degrades to a full real build and records an
     observable fell_through_reason. Back-fill writes are best-effort.
     WrapperError trimmed to its one genuine variant (couldn't spawn
     docker).

  2. u128 JSON round-trip: WrapperOutcome::CacheMiss carried a u128 ms
     duration; serde_json cannot deserialize a u128 through the buffer an
     internally-tagged (#[serde(tag="kind")]) enum requires, so every
     CacheMiss receipt — the dominant outcome — was un-round-trippable
     through the keyway "JSON receipt out" contract. Durations are now
     u64 ms (~584M years of range).

All non-docker tests pass; the #[ignore]d real-docker tests pass against
docker 28.5.2 + buildx 0.35.0 on arm64 (warmth: cold 4319ms, warm 280ms,
15.4x, 8/11 partial invalidation).

+++ sui-dockerfile-wrapper/tests/warmth_benchmark.rs
@@ -61,8 +61,8 @@ use sui_spec::dockerfile::{self, DockerfileArgs, MockDockerfileEnvironment};
 /// this harness produces (never a free-form log line).
 #[derive(Debug, Serialize)]
 struct WarmthReport {
-    cold_duration_ms: u128,
-    warm_duration_ms: u128,
+    cold_duration_ms: u64,
+    warm_duration_ms: u64,
     speedup_ratio: f64,
     partial_invalidation_node_count: usize,
     total_node_count: usize,
